+++ Source/Object/USceneComponent.h
@@ -1,5 +1,5 @@
 ﻿#pragma once
-#include "UObject.h"
+#include "Core/UObject/Object.h"
 #include "Core/Math/Vector.h"
 #include "Core/Container/Set.h"
 #include "Core/Math/Transform.h"
